sync fun init templates to initProject
